requirements.txt
langchain
langchain_community
langchain_core
langchain_groq
langchain_cohere
athina
voyageai
python-dotenv
tqdm
pandas
scikit-learn
openpyxl
datasets
wandb
pinecone_client
weave
rouge
athina-logger